Add prompt stack to fix bug where prompt is not restored after returning from shell-mode
When returning from shell-mode, the prompt is not properly restored.
The problem is that it gets clobbered by the new prompt, so we need somewhere to store it.
This adds a prompt stack so you can push and pop prompts in order to temporarily change
the current prompt and then later restore the previous prompt.

prompt and prompt= are also overridden to have meaning in the context of the prompt stack.