supercharged the `edit` command
(1) It now dumps contents of input buffer in an editor when you type 'edit' (or edit -t) inside a non-empty input buffer
(2) it accepts the file:line syntax. This closes #204

When it evals files it now (by default) just replaces the content of eval_string with the content of the file. This is particularly useful if content of file was an incomplete expression as no error is raised instead
the input buffer just awaits more input.