Rewrite 'edit-command'.
You no longer need to provide a variable that references a CommandSet…
Instead, you can provide just the command name.

The only cavaet is that the file defining the command you're editing
must return the CommandSet it is being written within.

You can also rewrite commands temporarily (in-memory) with the -p
switch.