Set _file_ and _line_ when calling whereami.
When a user would place a call to `binding.pry` followed by calling `whereami`
the locals _file_ and _line_ would not be set properly. This commit fixes this
by using the helper method `set_file_and_dir_locals()` (which also ensures _dir_
is set properly) as well as `_pry_.inject_local`.

Currently the _file_ local is set to /path/to/your/directory/(pry) if the
binding was set in a Pry session itself. I'm not entirely sure if this is
expected behavior, if this isn't the case it shouldn't be too hard to detect and
fix in a following commit.

This commit fixes #574.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>