Set Pry::Slop equal to Slop (v2.4.4)
We've decided not to upgrade our slop dependency to version 3 for
several reasons:

1. People who are currently using pry in their programs are using slop
version 2; and they may not realise this. Bumping the slop dependency
could therefore break people's code in subtle ways. By removing the
dependency completely we make it obvious that breakage has occurred, and
allows programs to chose their own version of slop.

2. The new version of slop is incompatible with pry. As a command-line
tool we use our option parsing library very very intensely; and as such
we've grown used to (and even embraced) it's warts, inconcistencies and
undocumented behaviours. Upgrading the slop dependency at this point
requires a lot of fix up work both on the slop codebase and the pry
codebase.

3. The slop API is implicitly part of the pry API for defining custom
commands; by changing that API we run a significant risk of breaking
people's pry extensions.

4. The problems in 1., 2. and 3. are going to recur every time slop releases
a new version.

5. Doing this does not preclude upgrading the version of slop we use in
the future. Beacuse 1. will not occur anymore, the only breakage that we
can cause is in pry extensions; this seems resonable providing we signal
the fact that we're doing it (for example by updating the major
version number).