Change 'pry-version' command to 'version'
I did some research on this topic and it turns out that initially Pry
had `version` command instead of `pry-version`. Since v0.9.3 Pry
started to use `pry-version` command. Since then, in order to know the
version of Pry users have to use `pry-version` command, whereas
`<= 0.9.2` users use `version`.

Research
--------

Let's ensure, that both of these commands really work with the specific
version of Pry.

* Pry >= 0.9.3:

    % pry --simple-prompt
    >> pry-version
    Pry version: 0.9.9.6 on Ruby 1.9.3.

* Pry <= 0.9.2:

    % pry --simple-prompt
    >> version
    Pry version: 0.9.2 on Ruby 1.9.3.

There is one problem in the new (pry-version) command: it's not obvious
to guess. I'll explain that hereinafter. Let's investigate the
situation further.

* Pry <= 0.9.2:

    % pry --simple-prompt
    >> version
    Pry version: 0.9.2 on Ruby 1.9.3.
    >> pry-version
    >> ^D
    NoMethodError: undefined method `-' for main:Object
    from (pry):in `<main>'

* Pry >= 0.9.3:

    % pry --simple-prompt
    >> pry-version
    Pry version: 0.9.9.6 on Ruby 1.9.3.
    >> version
    => ">= 0"
    >> help version
    pry-version        Show Pry version.

Wait, what? `version` still works? But how the hell?! Definitely
something is wrong up here. The autocompletion works for `version`,
which adds even more confusion to the user:

    >> ver<TAB>
    >> version

Well, the answer is that `version` is a variable:

    >> defined? version
    => "local-variable"

The Answer
----------

The variable `version` is defined in `which pry` executable. So when
we call `version`, we don't execute Pry's command, we call the variable
from RubyGems.

Wrap Up
-------

I do think, this is a problem. Let me explain my opinion. First of all,
`pry-version` and `pry-backtrace` are the only two commands that use
`pry-*` prefix. But I don't see any logical explanation why `version`
has been changed to `pry-version` (the commit in the beginning of the
issue doesn't explain that). Why not to use `pry-*` prefix for every
single command, then? `pry-nesting`, `pry-cat`, `pry-play` and so on...
Because it's verbose! So is `pry-version`.

At first, I thought it's a bug in Pry (I forgot that `pry-version`
command exists, and thought that `version` is the only valid command
for that action, but that's my individual problem). I've spent some
time, trying to figure out, what was that "version" thingie and where
it comes from, but the output from `help` confused me even more. I
wonder, why does RubyGems define the `version` variable at the
top-level of your application every time when you use executables?

My solution of the problem is to change `pry-version` back to `version`.

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>