Refactor `--installed-plugins` switch
First of all, add `PluginManager.show_installed_plugins` method. The
body of this method is taken from `--installed-plugins` code, which
means that it still displays the list of all installed plugins.

Secondly, use `PluginManager.show_installed_plugins` in
`--installed-plugins` body instead of the old code.

This change is needed, because we want to invoke `--installed-plugins`
from other place of the program and we don't want to repeat ourselves.

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>