Improve `-I` option support
Namely, add support for multiple arguments for the `-I` option. For
example:

  % pry -Ilib:../code/src

Where `:` is the delimiter for paths. In the example above `./lib` and
`../code/src` directories will be added to $LOAD_PATH array.

Also, there was a bug in Slop with respect to parsing of "fuzzy options
with arguments"[1]. It is fixed now, so the next version of Slop will
indirectly resolve our issue #674 (-r/--require doesn't behave like
irb's)[2].

And finally, wrap some very long lines in the "cli.rb" (this is just a
cosmetic change).

[1]: https://github.com/injekt/slop/issues/74
[2]: https://github.com/pry/pry/issues/674

Signed-off-by: Kyrylo Silin <kyrylosilin@gmail.com>