Eliminate spurious hashrockets.
These were coming from a pretty simple source: the DEFAULT_PRINT
prepends them so you can get output like:

[1] pry(main)> 2+3
=> 5

But since we're formatting these differently, obviously we don't want
that prefix. So this patch extracts a Pry.format_for_output method then
calls that iff the user hasn't changed the default.

There is some hackitude involved, but the test pass (and are of decent
coverage, I think, so feel fry to try to diff this down if you have a
good idea.)