Allow pry to be disabled [Fixes #497]
It's easy to find yourself in 'pry-hell' when you've added a binding.pry in a
tight loop, or if you've got pry-rescue or plymouth enabled and you've broken
hundreds of tests. To get out of this, the user can just type 'disable-pry'.

If you suspect in advance that pry is going to be unhelpful you can temporarily
disable it without having to edit the source code of your program by exporting
NO_PRY=true.