Improve Ruby 2.0 refinement support, closes #738
added lookup_method_via_binding(), using this rather than safe_send
we respect the method object introduced by any refinements in the binding.

Further work can be done for refinement support -- such as indicating that the method is a refinement, showing
the name of the refining module, and so on.