Use correct_indent not auto_indent...
Prior to this change, when auto_indent was set to true for some tests,
the green dots would be cleared half-way through.

We should probably re-consider these options, but it seems useful to
explicitly disable ANSI code rewriting during specs while leaving
indentation mutating read lines (before this change, there was a bug
in REPL that did not manifest because auto_indent was false).