Robustify the plugin require
I was getting trouble with both pry-editline and pry-syntax-hacks, while
running with a newly-compiled-Ruby with GNU Readline on a Mac (the
default Readline was working OK with these. There could be more
investigation about the error, but for now, at least pry boots with a
handy error instead of a total failure).