ShowDocSpec: add another random comment to the definition
This commit is rather interesting. Well, maybe not the commit, but the
issue it is aimed to fix. So what's the aim? The aim is to eliminate an
error in the test for the following Ruby implementations: rbx-18mode,
jruby-18mode, jruby-19mode.

See that build for details: https://travis-ci.org/pry/pry/builds/5478834

You may be wondering now: "Kyrylo, what is so interesting in it?".
Heh-heh, sit down right beside me, my covey.

I have a screenshot for you:
http://img-fotki.yandex.ru/get/4125/98991937.11/0_911d2_acdd6798_orig

The screenshot demonstrates the difference between `show-doc` finding
monkey-patches on different Ruby implementations (JRuby and MRI 1.9.3).
Take a closer look at the order of the candidates. It is different in
the abovementioned implementations. That is the reason why the test is
failing. Unfortunately, I couldn't figure out why that is happening.

The fix is simple and it doesn't ruin the idea of the test. However,
I'm leaving that order issue to the greater minds of our civilization.

Uh! Don't forget to check out the description of that commit, because it
is related to this topic:
https://github.com/pry/pry/commit/e926d8c9444d025c5d86489aea0111b54378f396