Remove man page from pry
This removal was requested because the *nix man command does not find
it.  The pry man page was used by a rarely used gem named gem-man, that
looked in a man folder for the correct page.

Until rubygems plan to load manpages into GEM_PATH for general use

https://github.com/rubygems/rubygems/issues/74

We'll remove this.