lib/pry/commands/ls.rb: fix `ls -c` on top-level
This is rather an odd bug introduced by this commit: "Add support for
BasicObjects to `ls`"[1]. `ls -c` works, if you install it as a gem, but
it refuses to display anything, if you run it via `rake pry`. Moreover,
I've tried to write a failing test without changing the implementation,
but the test suite couldn't fail. Let's fix this issue, because it can
save us a lot of time, avoiding the situation when one stumble upon this
oddity once again.

[1]: ac683892d10bc88919e51604da5f80e9c706720b