update the process uses to set the (default) configuration options.
the class Pry defines a bunch of different configuration options that
change the behavior of a pry session once it begins. these options are
inherited by an instance of Pry who implements a REPL loop. the default
options can be changed by modifying the Pry instance(available as '_pry_' in
the REPL).

these changes make it easier to add new default options by adding a key & value
pair to Pry::CONFIG. the key should be the name of the option and the value should
be a Proc that reads a value from the Pry class.

this should help solve #1055.