fix the failing Pry#prompt specs.
this is a strange and odd case. Pry.prompt is a delegate to Pry.config, as
it has always been. the same delegate was setup on an instance of Pry, but
never used because we define #prompt and #prompt= with our implementation.

the thing that would make the most sense (to me) is to not support Pry.prompt
anymore and recommend the use of Pry.config.prompt instead. a lot of code relies
on Pry.prompt though, so we have to support the delegate to config and implement
custom behavior for the pry instance.