add bond as a runtime dependency.
bond offers some cool features in the realm of completion but most users
will fallback onto `Pry::InputCompleter` that doesn't seem to provide the
same features or extensibility.

pry will use bond if:
a) user installs bond gem separately
b) user requires "bond" in .pryrc
c) if user is on OSX, user shouldn't be using editline (OSX default)

c) stays true with this commit because it seems bond doesn't work very
well with it, but if you don't have editline or you aren't on OSX bond
is assumed to be the default.

the disadvantage is that we're adding another runtime dependency to pry.
i think the advantage is that users can take advantage of better completion
in pry by default and pry could offer some other cool completions, like bond-yard:
https://github.com/cldwalker/bond-yard

atm the defaults make it unlikely bond will ever be used.
maybe close #1160