add 'list-inspectors' & 'change-inspectors' commands.
the list-inspectors command shows a list of available inspectors as well
as the active inspectors in the list. inspectors are drawn with a name as
well as a short description.

change-inspector can be used to change the inspector by the name found in
the list-inspectors command.

an inspector in pry is something that prints a return value in a repl session.
i'm not crazy about the name `inspector` but the traditional name used in pry(`print`)
didn't fit well. some discussion led us to `inspector` for lack of a better word,
see #1176 on github for discussion