Fix requiring of rubygems
The Rubygem documentation here http://guides.rubygems.org/name-your-gem/
suggests the naming of rubygems. If the name contains dashes (-) the
require statements should include forward slashes that replace the dashes.

For example, the 'net-ssh' rubygem should have the require statement of
```ruby
require 'net/ssh'
```

This simple fix tries to address this problem and falls back to the
original require statement if there is a `LoadError`. This will handle
cases where some gem author didn't follow the naming convention
suggested in the documentation.