Fix delegation of `quiet?` config method (fix #1261)
Because we were defining `quiet?` on Pry::Config::Default, calls to it
would always delegate to the default config object and then return
false. To make it possible to enable quiet mode, we need the
transformation from `quiet?` to `quiet` to happen earlier. The simplest
way to accomplish this is just to define the `quiet?` method explicitly.