Use CPU-less platforms for windows
There is no need to include the CPU in the platform string as there is
no compiled code shipped in the pry gems.  The only difference is the
dependency.

Pry became a victim of rubygems/rubygems#1058 which will be fixed in
rubygems/rubygems#1120.  I would also like to raise awareness of proper
use of platform parts while I'm fixing the bug in RubyGems.