extend Pry::Prompt for integrating custom prompts.
A prompt added through Pry::Prompt will be visible in the output of
the "list-prompts" command, and it can be used with the "change-prompt"
command. So, adding a prompt this way integrates with the rest of Pry
and is a good option for plugins who want their prompts to be discoverable
by Pry's built-in commands.