workaround JRuby CI issue.
'jruby' as a catch-all often fails (not always) but pinning to a specific
version seems to at least be able to run the test suite. Tracking JRuby
development may be best, and then pinning to a new version of JRuby when
a new version is released. It's not ideal, but at least it doesn't cause
false red flags.