add gem-stat command
thanks to @dannytatom for the original gem-stats rubygems plugin, which
inspired this Pry command.

it returns some general statistics about a rubygem when given a name.
note that rubygems rate limits the requests, and thanks to that this
command can be a "Fail Whale" although the error is handled gracefully.

This commit depends on pull request:
https://github.com/pry/pry/pull/1701