The following methods take a third required argument, an instance of Pry,
in order to fulfil queries to `_pry_.config`.

* Pry::Helpers.tablify
* Pry::Helpers.tablify\_to\_screen\_width
* Pry::Helpers.tablify\_or\_one\_line

The "Pry::Helpers::Table" class also takes the same required third argument.
**Breaking change**

Although a breaking change, it brings this code in line with how the rest of Pry
works, where session-local configuration queries are made to `_pry_.config` and
not `Pry.config` (which sets defaults, normally from ~/.pryrc or plugin code)