Add Pry#h.
Returns a Module that provides methods normally available to Pry commands and who
are also aware of this Pry instance. Some helper modules have been excluded because
I am not sure how useful they are, yet.

A real case example of when this method is useful:
https://github.com/r-obert/pry-larryify/blob/7c920d9981fea51637d3c13e932f4253cdf4a850/lib/pry-larryify.rb#L79

That line can become simply:
:LarrySays => pry.h.green("Benchmark:")

Because `pry.h` is aware of the Pry instance it is associated with,
as are the methods implemented on top of it. This also makes the pry
instance that is passed around more powerful, and lends itself well
to a replacement for scenarios where 'Pry::Helpers::X.singleton_method?'
is used directly.

See '@example' in the documentation for this method as well.

The Pry::Deprecate pull request will depend on this commit, eg:
pry.h.deprecate_method! [method(:foo)], "foo is deprecated, use #bar instead."

Unfortunately other parts of Pry are coupled to always writing colour, so we
cannot remove Output#decolorize_maybe until. This PR opens the path for 'Pry#h'
though.