PluginManager#load_cli_options: use the realpath
Since ruby 2.5, `require 'foo'` will use the realpath of the file
corresponding to foo.rb. However `require '/absolute/path/to/foo.rb'`
won't use the realpath.

So when $GEM_HOME contains a symlink (ie it is not the realpath), when a
pry plugin is loaded, by `activate!`:
    require gem_name if !active?
the real_path of `gem_name` is used.

But then in load_cli_options:
    cli_options_file = File.join(spec.full_gem_path, "lib/#{spec.name}/cli.rb")
    require cli_options_file if File.exist?(cli_options_file)
since the path given is absolute, it will be used directly without realpath.

This means that cli.rb may potentially be required twice (once via its realpath if the plugin requires it, the other via its $GEM_HOME path when required by `load_cli_options`), which could raise some errors.

Fix this by using the realpath in load_cli_options too.