plugins: use require_relative instead of require
I've stumbled upon the following commit:
https://bugs.ruby-lang.org/projects/ruby-trunk/repository/revisions/63046/diff/spec/ruby/core/file/ftype_spec.rb

It looks like `require_relative` resolves symlinks for us, so there's no need to
use `realpath` introduced in https://github.com/pry/pry/pull/1762.

One caveat is that `require_relative` is not a thing in Ruby 1.9, however the
plan is to drop support for it for
v1.0.0 (https://github.com/pry/pry/issues/1371).

This is more of a proof of concept for now since we *still* support 1.9.3, so
it cannot be merged immediately.