repl: correctly calculate overhang for empty lines
This improves on https://github.com/pry/pry/pull/1813.

I spotted a problem with the way it works: sometimes, if you enter an empty
line, the output would span across two lines:

```
[1] pry(main)>

[2] pry(main)>

[3] pry(main)>
```

This is because overhang is bigger than the current line. I also noticed a bug
with method definition, where defining a method leaves an unwanted empty line:

```
[3] pry(main)> def foo
             | end

=> :foo
```

To fix this I changed the way we calculate overhang. First and the most
important change is that I stopped calculating overhang for emacs mode. This is
because it's too risky to introduce this change because it's the default mode
and 99% of our users use this, so there's no need to change this
behaviour. Another reason is that emacs mode users typically don't use any mode
indicators (because emacs has no modes), so it strikes me as a more pragmatic
solution.

With Vi mode we calculate overhang and still support custom indicators.