pager.rb: fix Helpers constant invocation
Since the top class is Pry::Pager and not Pry, the cref constant static lookup
only look for constants under Pry::Pager (or at the top level).

Since Helpers refers to Pry::Helpers, it won't be found. Fix this by
changing the top class to Pry and put the class Pager in the Pry
namespace.

This introduce a two space shift of all the code, so reformat the
comments.

This also fixes the following bug: in SystemPager.available? the Helpers
constant lookup fails, so this raises an exception which is rescued and
then returns false; so the SimplePager is always chosen instead.