Update Pry::Config::Behavior#eager_load!
I don't think `#eager_load!` was working properly on Pry master or
even any of the released Prys from a quick test I did.

The `memoized_methods` method `eager_load!` is relying on has
since been removed from master too, so an update is needed.

The new implementation is simpler, and it includes a test case to
check that its behaving as expected because my manual tests suggest
it was not (I observed some keys being loaded but not all of them).