Factor out gem-* commands to a separate plugin
This code doesn't really belong to Pry Core. It is all about RubyGems, but Pry
is all about introspection and debugging. We used to have some gem-* commands in
the past and it wasn't a big issue but now, since we have 7 commands, they
really deserve their own place.

These commands were moved to:
https://github.com/pry/pry-gem