Delete the `gist` command
This command was moved to the [pry-jist][1] plugin. Pry allows good
extensibility and this command doesn't feel like something that should be in Pry
Core. I suspect not many people use it, let alone know about it.

The concept of optional dependencies also feels awkward (the `gist` gem). It's a
strong indication that the command should be moved outside Pry Core (or we
should just depend on `gist`).

[1]: https://github.com/pry/pry-jist